+++ zingolib/CHANGELOG.md
@@ -64,3 +64,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `do_save_to_buffer`
   - `do_save_to_buffer_sync`
   - `fix_spent_at_height`
+  - `TransactionRecord::net_spent`
+  - `TransactionRecord::get_transparent_value_spent()`